Re-compute CRC32 of converted movesets, fix conversion of compressed movesets

+++ TKMovesets2/Utils/Compression.cpp
@@ -476,11 +476,17 @@ namespace CompressionUtils
 	{
 		namespace Moveset
 		{
+			bool Compress(const std::wstring& filename, TKMovesetCompressionType_ compressionType)
+			{
+				return Compress(filename, filename, compressionType);
+			}
+
 			bool Compress(const std::wstring& dest_filename, const std::wstring& src_filename, TKMovesetCompressionType_ compressionType)
 			{
 				std::ifstream orig_file(src_filename, std::ios::binary);
-				std::ofstream new_file(dest_filename, std::ios::binary);
 
+				uint32_t moveset_data_start;
+				char* moveset_header_buf;
 				// Copy up to moveset_data_start, get size of moveset data
 				uint64_t moveset_data_size;
 				{
@@ -489,36 +495,52 @@ namespace CompressionUtils
 					orig_file.seekg(0, std::ios::beg);
 					orig_file.read((char*)&header, sizeof(header));
 
-					uint32_t moveset_data_start = header.moveset_data_start;
+					moveset_data_start = header.moveset_data_start;
 
 					// Read up to moveset data start (includes header and more)
-					char* buf = new char[moveset_data_start];
+					moveset_header_buf = new char[moveset_data_start];
 					orig_file.seekg(0, std::ios::beg);
-					orig_file.read(buf, moveset_data_start);
+					orig_file.read(moveset_header_buf, moveset_data_start);
 
 					// Calculate moveset data size
 					orig_file.seekg(0, std::ios::end);
 					moveset_data_size = (uint64_t)orig_file.tellg() - moveset_data_start;
 
 					// Mark moveset as compressed
-					((TKMovesetHeader*)buf)->compressionType = compressionType;
-					((TKMovesetHeader*)buf)->moveset_data_size = moveset_data_size;
-
-					// Write header
-					new_file.write(buf, moveset_data_start);
-					delete[] buf;
+					((TKMovesetHeader*)moveset_header_buf)->compressionType = compressionType;
+					((TKMovesetHeader*)moveset_header_buf)->moveset_data_size = moveset_data_size;
 
 					// Move cursor back to the start of the moveset data
 					orig_file.seekg(moveset_data_start, std::ios::beg);
 				}
 
 				// Compress moveset data
-				Byte* inbuf = new Byte[moveset_data_size];
+				Byte* inbuf;
+				try {
+					inbuf = new Byte[moveset_data_size];
+				}
+				catch (std::bad_alloc& ex) {
+					DEBUG_ERR("Compress: Failed to allocate %llu bytes (%s)", moveset_data_size, ex.what());
+					return false;
+				}
+
 				Byte* outbuf = nullptr;
 				orig_file.read((char*)inbuf, moveset_data_size);
 				orig_file.close();
 
-				std::filesystem::remove(src_filename);
+				// Write header
+				std::ofstream new_file(dest_filename, std::ios::binary);
+				new_file.write(moveset_header_buf, moveset_data_start);
+				delete[] moveset_header_buf;
+
+				if (src_filename != dest_filename) {
+					try {
+						std::filesystem::remove(src_filename);
+					}
+					catch (std::exception& ex) {
+						DEBUG_ERR("Compression: (%s)", ex.what());
+					}
+				}
 				uint64_t compressed_size = 0;
 
 				switch (compressionType)
